requirements.txt
aiogram==2.23.1
aiogram-calendar==0.1.1
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.68.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.1
googleapis-common-protos==1.57.0
httplib2==0.21.0
idna==3.4
magic-filter==1.0.9
multidict==6.0.3
oauthlib==3.2.2
protobuf==4.21.10
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.13
yarl==1.8.2
